refactor(filehost): add useFileHost hook
